+++ SupermarketManagement/src/main/java/com/ou/repositories/MemberTypeRepositoryForTest.java
@@ -1,6 +1,5 @@
 package com.ou.repositories;
 
-import com.ou.pojos.Member;
 import com.ou.pojos.MemberType;
 import com.ou.pojos.Sale;
 import com.ou.pojos.SalePercent;
@@ -91,7 +90,7 @@ public class MemberTypeRepositoryForTest {
             if(resultSet.next()){
                 salePercent.setSperId(resultSet.getInt("sper_id"));
                 salePercent.setSperIsActive(resultSet.getBoolean("sper_is_active"));
-                salePercent.setSperPercent(resultSet.getFloat("sper_percent"));
+                salePercent.setSperPercent(resultSet.getInt("sper_percent"));
             }
         } catch (SQLException e) {
             e.printStackTrace();
