+++ SupermarketManagement/src/main/java/com/ou/repositories/SaleRepository.java
@@ -51,7 +51,7 @@ public class SaleRepository {
                 SalePercent salePercent = new SalePercent();
                 salePercent.setSperId(resultSet.getInt("sper_id"));
                 salePercent.setSperIsActive(resultSet.getBoolean("sper_is_active"));
-                salePercent.setSperPercent(resultSet.getFloat("sper_percent"));
+                salePercent.setSperPercent(resultSet.getInt("sper_percent"));
                 sale.setSalePercent(salePercent);
                 sales.add(sale);
             }
